Move guards to separated folder
